+++ react-client/src/components/chatbotAutomation/header.js
@@ -6,17 +6,11 @@ class Header extends React.Component {
   constructor(props, context) {
     super(props, context)
     this.state = {
-      waitingForPublish: false,
-      waitingForDisable: false,
       title: props.title,
       editTitle: false
     }
     this.onDelete = this.onDelete.bind(this)
     this.afterDelete = this.afterDelete.bind(this)
-    this.onDisable = this.onDisable.bind(this)
-    this.afterDisable = this.afterDisable.bind(this)
-    this.onPublish = this.onPublish.bind(this)
-    this.afterPublish = this.afterPublish.bind(this)
     this.onEditTitle = this.onEditTitle.bind(this)
     this.onCancelEdit = this.onCancelEdit.bind(this)
     this.onRename = this.onRename.bind(this)
@@ -33,34 +27,6 @@ class Header extends React.Component {
     } else {
       this.props.alertMsg.error(res.description)
     }
-  }
-
-  onDisable () {
-    this.setState({waitingForDisable: true})
-    this.props.onDisable(this.afterDisable)
-  }
-
-  afterDisable (res) {
-    if (res.status === 'success') {
-      this.props.alertMsg.success('Bot disbabled successfully')
-    } else {
-      this.props.alertMsg.error(res.description)
-    }
-    this.setState({waitingForDisable: false})
-  }
-
-  onPublish () {
-    this.setState({waitingForPublish: true})
-    this.props.onPublish(this.afterPublish)
-  }
-
-  afterPublish (res) {
-    if (res.status === 'success') {
-      this.props.alertMsg.success('Bot published successfully')
-    } else {
-      this.props.alertMsg.error(res.description)
-    }
-    this.setState({waitingForPublish: false})
   }
 
   onEditTitle () {
@@ -135,81 +101,6 @@ class Header extends React.Component {
             </h3>
           }
         </div>
-        <div className='col-md-6'>
-          {
-            this.props.isPublished
-            ? <button
-              id='_chatbot_message_area_header_disable'
-              style={{marginLeft: '10px'}}
-              type='button'
-              className={`pull-right btn btn-danger m-btn m-btn--icon ${this.state.waitingForDisable && 'm-loader m-loader--light m-loader--left'}`}
-              onClick={this.onDisable}
-            >
-              <span>
-                {
-                  !this.state.waitingForDisable &&
-                  <i className='fa flaticon-close' />
-                }
-                <span>Disable</span>
-              </span>
-            </button>
-            : <button
-              id='_chatbot_message_area_header_publish'
-              style={{marginLeft: '10px'}}
-              type='button'
-              className={`pull-right btn btn-success m-btn m-btn--icon ${this.state.waitingForPublish && 'm-loader m-loader--light m-loader--left'}`}
-              onClick={this.onPublish}
-              disabled={!this.props.canPublish}
-            >
-              <span>
-                {
-                  !this.state.waitingForPublish &&
-                  <i className='fa flaticon-paper-plane' />
-                }
-                <span>Publish</span>
-              </span>
-            </button>
-          }
-          <button
-            id='_chatbot_message_area_header_test'
-            style={{marginLeft: '10px'}}
-            type='button'
-            className='btn btn-primary pull-right m-btn m-btn--icon'
-            onClick={this.props.onTest}
-            disabled={!this.props.canTest}
-          >
-            <span>
-              <i className='fa flaticon-chat-1' />
-              <span>Test</span>
-            </span>
-          </button>
-          {this.props.showAnalytics &&
-          <button
-            id='_chatbot_message_area_header_analytics'
-            style={{marginLeft: '10px'}}
-            type='button'
-            className='btn btn-info pull-right m-btn m-btn--icon'
-            onClick={this.props.onAnalytics}
-          >
-            <span>
-              <i className='fa flaticon-analytics' />
-              <span>Analytics</span>
-            </span>
-          </button>
-        }
-          {
-            this.props.showDelete &&
-            <button
-              id='_chatbot_message_area_header_delete'
-              style={{marginLeft: '10px'}}
-              type='button'
-              className='pull-right btn btn-primary'
-              onClick={() => this.refs._delete_message_block.click()}
-            >
-              Delete Step
-            </button>
-          }
-        </div>
         <button style={{display: 'none'}} ref='_delete_message_block' data-toggle='modal' data-target='#_cb_ma_delete_mb' />
         <CONFIRMATIONMODAL
           id='_cb_ma_delete_mb'
@@ -226,12 +117,6 @@ Header.propTypes = {
   'title': PropTypes.string.isRequired,
   'showDelete': PropTypes.bool.isRequired,
   'onDelete': PropTypes.func.isRequired,
-  'onTest': PropTypes.func.isRequired,
-  'canTest': PropTypes.bool.isRequired,
-  'canPublish': PropTypes.bool.isRequired,
-  'onPublish': PropTypes.func.isRequired,
-  'onDisable': PropTypes.func.isRequired,
-  'isPublished': PropTypes.bool.isRequired,
   'blocks': PropTypes.array.isRequired
 }
