+++ react-client/src/components/chatbotAutomation/sidebar.js
@@ -224,8 +224,8 @@ class Sidebar extends React.Component {
   render () {
     return (
       <div id='_chatbot_sidebar' style={{border: '1px solid #ccc', backgroundColor: 'white', padding: '0px'}} className='col-md-3'>
-        <div style={{margin: '0px'}} className='m-portlet m-portlet-mobile'>
-          <div id='_chatbot_sidebar_items' style={{height: '80vh'}} className='m-portlet__body'>
+        <div style={{margin: '0px', height: '100%'}} className='m-portlet m-portlet-mobile'>
+          <div id='_chatbot_sidebar_items' style={{padding: '1.2rem 1.2rem', overflowY: 'scroll'}} className='m-portlet__body'>
             {
               this.state.expandedNodes.length > 0 &&
               <TreeView
@@ -243,43 +243,43 @@ class Sidebar extends React.Component {
               restoreBackup={this.restoreBackup}
               toggleBackupModal={this.toggleBackupModal}
             />
-            <div
-              style={{
-                position: 'absolute',
-                bottom: 0,
-                left: 0,
-                width: '100%',
-                paddinng: '1.1rem',
-                borderTop: 'none'
-              }}
-              className="m-portlet__foot"
-            >
-  						<div className="row align-items-center">
-  							<div className="col-lg-12">
-                  <button
-                    ref='chatbotBackup'
-                    style={{display: 'none'}}
-                    data-toggle='modal'
-                    data-target='#_chatbot_backup_modal'
-                  />
-                  <button
-                    style={{
-                      background: 'none',
-                      border: 'none',
-                      float: 'right',
-                      cursor: 'pointer'
-                    }}
-                    className="m-link m--font-boldest m-btn m-btn--icon"
-                    onClick={this.openBackupModal}
-                  >
-                    <span>
-                      <i className='fa fa-cloud-upload'/>
-                      <span>Backup</span>
-                    </span>
-                  </button>
-  							</div>
-  						</div>
-  					</div>
+          </div>
+          <div
+            style={{
+              position: 'absolute',
+              bottom: 0,
+              left: 0,
+              width: '100%',
+              paddinng: '1.1rem',
+              borderTop: 'none'
+            }}
+            className="m-portlet__foot"
+          >
+            <div className="row align-items-center">
+              <div className="col-lg-12">
+                <button
+                  ref='chatbotBackup'
+                  style={{display: 'none'}}
+                  data-toggle='modal'
+                  data-target='#_chatbot_backup_modal'
+                />
+                <button
+                  style={{
+                    background: 'none',
+                    border: 'none',
+                    float: 'right',
+                    cursor: 'pointer'
+                  }}
+                  className="m-link m--font-boldest m-btn m-btn--icon"
+                  onClick={this.openBackupModal}
+                >
+                  <span>
+                    <i className='fa fa-cloud-upload'/>
+                    <span>Backup</span>
+                  </span>
+                </button>
+              </div>
+            </div>
           </div>
         </div>
       </div>
