+++ react-client/src/components/chatbotAutomationNew/messageArea.js
@@ -483,15 +483,11 @@ MessageArea.propTypes = {
   'chatbot': PropTypes.object.isRequired,
   'uploadAttachment': PropTypes.func.isRequired,
   'handleAttachment': PropTypes.func.isRequired,
-  'currentLevel': PropTypes.number.isRequired,
-  'maxLevel': PropTypes.number.isRequired,
   'blocks': PropTypes.array.isRequired,
   'sidebarItems': PropTypes.array.isRequired,
   'handleMessageBlock': PropTypes.func.isRequired,
-  'fbAppId': PropTypes.string.isRequired,
   'changeActiveStatus': PropTypes.func.isRequired,
   'deleteMessageBlock': PropTypes.func.isRequired,
-  'registerSocketAction': PropTypes.func.isRequired,
   'progress': PropTypes.number.isRequired,
   'updateParentState': PropTypes.func.isRequired,
   'checkWhitelistedDomains': PropTypes.func.isRequired,
