+++ react-client/src/components/sidebar/sidebar.js
@@ -182,18 +182,11 @@ class Sidebar extends Component {
           route: '/chatbots'
         })
       }
-      if (platform === 'messenger' && (isKiboChat || isLocalhost) && user.isSuperUser && user.plan['chatbot_automation'] && user.permissions['configure_chatbot_automation']) {
-        submenu.push({
-          priority: 'b',
-          name: 'Chatbot Automation (OLD)',
-          route: '/chatbotAutomation'
-        })
-      }
       if (platform === 'messenger' && (isKiboChat || isLocalhost) && user.plan['chatbot_automation'] && user.permissions['configure_chatbot_automation']) {
         submenu.push({
           priority: 'b',
           name: 'Chatbot Automation',
-          route: '/chatbotAutomationNew'
+          route: '/chatbotAutomation'
         })
       }
       if ((isKiboEngage || isLocalhost) && user.plan['autoposting'] && user.permissions['view_autoposting_feeds']) {
