+++ react-client/src/containers/chatbotAutomation/configureChatbot.js
@@ -19,14 +19,20 @@ import { saveWhiteListDomains, fetchWhiteListedDomains, deleteDomain } from '../
 import { registerAction } from '../../utility/socketio'
 import { Prompt } from 'react-router'
 import AlertContainer from 'react-alert'
+import $ from 'jquery'
+import ReactTooltip from 'react-tooltip'
+
+import ADVANCEDBUILDER from './builder/advancedBuilder'
+import BASICBUILDER from './builder/basicBuilder'
 import PROGRESS from '../../components/chatbotAutomation/progress'
-import SIDEBAR from '../../components/chatbotAutomation/sidebar'
-import MESSAGEAREA from '../../components/chatbotAutomation/messageArea'
 import BACKBUTTON from '../../components/extras/backButton'
 import HELPWIDGET from '../../components/extras/helpWidget'
 import MODAL from '../../components/extras/modal'
+import CONFIRMATIONMODAL from '../../components/extras/confirmationModal'
 import WHITELISTDOMAINS from '../../components/chatbotAutomation/whitelistDomains'
-import $ from 'jquery'
+import DROPDOWN from '../../components/extras/dropdown'
+
+const MessengerPlugin = require('react-messenger-plugin').default
 
 class ConfigureChatbot extends React.Component {
   constructor (props, context) {
@@ -37,12 +43,17 @@ class ConfigureChatbot extends React.Component {
       blocks: [],
       sidebarItems: [],
       currentBlock: {title: ''},
-      currentLevel: 1,
       progress: 0,
       showWhitelistDomains: false,
       unsavedChanges: false,
       attachmentUploading: false,
-      allTriggers: []
+      allTriggers: [],
+      powerLoading: false,
+      showTestContent: false,
+      existingChatbot: props.location.state.existingChatbot,
+      showTreeStructure: false,
+      builder: props.location.state.chatbot.builderPreference,
+      builderText: 'Basic Builder'
     }
 
     this.fetchChatbotDetails = this.fetchChatbotDetails.bind(this)
@@ -58,15 +69,19 @@ class ConfigureChatbot extends React.Component {
     this.toggleWhitelistModal = this.toggleWhitelistModal.bind(this)
     this.getWhitelistModalContent = this.getWhitelistModalContent.bind(this)
     this.onAnalytics = this.onAnalytics.bind(this)
+    this.onDisable = this.onDisable.bind(this)
+    this.afterDisable = this.afterDisable.bind(this)
+    this.onPublish = this.onPublish.bind(this)
+    this.afterPublish = this.afterPublish.bind(this)
+    this.showTestModal = this.showTestModal.bind(this)
+    this.toggleTestModalContent = this.toggleTestModalContent.bind(this)
+    this.getTestModalContent = this.getTestModalContent.bind(this)
+    this.showTreeStructure = this.showTreeStructure.bind(this)
+    this.closeTreeStructure = this.closeTreeStructure.bind(this)
+    this.disableShopifyChatbot = this.disableShopifyChatbot.bind(this)
+    this.switchBuilder = this.switchBuilder.bind(this)
 
     props.getFbAppId()
-  }
-
-  onAnalytics () {
-    this.props.history.push({
-      pathname: '/chatbotAnalytics',
-      state: {chatbot: this.state.chatbot, page: this.props.location.state.page, backUrl: '/configureChatbot'}
-    })
   }
 
   componentDidMount () {
@@ -83,20 +98,16 @@ class ConfigureChatbot extends React.Component {
     /* eslint-disable */
     $('#sidebarDiv').addClass('hideSideBar')
     /* eslint-enable */
-  }
 
-  updateState (state) {
-    this.setState(state)
-  }
-
-  fetchChatbot () {
-    this.props.fetchChatbot(this.state.chatbot._id, this.handleChatbot)
-  }
-
-  handleChatbot (res) {
-    if (res.status === 'success') {
-      this.setState({chatbot: res.payload})
-    }
+    let comp = this
+    registerAction({
+      event: 'chatbot.test.message',
+      action: function (data) {
+        comp.msg.success('Sent successfully on messenger')
+        comp.refs._open_test_chatbot_modal.click()
+        comp.toggleTestModalContent()
+      }
+    })
   }
 
   fetchChatbotDetails () {
@@ -195,6 +206,10 @@ class ConfigureChatbot extends React.Component {
     }
   }
 
+  updateState (state) {
+    this.setState(state)
+  }
+
   getAllBlocks (data) {
     let blocks = data
     for (let i = 0; i < data.length; i++) {
@@ -224,6 +239,16 @@ class ConfigureChatbot extends React.Component {
     })
   }
 
+  fetchChatbot () {
+    this.props.fetchChatbot(this.state.chatbot._id, this.handleChatbot)
+  }
+
+  handleChatbot (res) {
+    if (res.status === 'success') {
+      this.setState({chatbot: res.payload})
+    }
+  }
+
   toggleWhitelistModal () {
     this.setState({showWhitelistDomains: true})
   }
@@ -242,6 +267,126 @@ class ConfigureChatbot extends React.Component {
     } else {
       return (<div />)
     }
+  }
+
+  onAnalytics () {
+    this.props.history.push({
+      pathname: '/chatbotAnalytics',
+      state: {chatbot: this.state.chatbot, page: this.props.location.state.page, backUrl: '/configureChatbot'}
+    })
+  }
+
+  onDisable() {
+    this.setState({ powerLoading: true })
+    const data = {
+      chatbotId: this.state.chatbot._id,
+      published: false
+    }
+    this.props.updateChatbot(data, (res) => this.afterDisable(res))
+  }
+
+  afterDisable(res) {
+    if (res.status === 'success') {
+      this.msg.success('Chatbot disabled successfully')
+      let chatbot = this.state.chatbot
+      chatbot.published = false
+      this.setState({ chatbot, powerLoading: false })
+    } else {
+      this.msg.error(res.description)
+    }
+  }
+
+  onPublish() {
+    if (this.state.existingChatbot && this.state.existingChatbot.published) {
+      this.disableShopifyChatbotTrigger.click()
+    } else {
+      this.setState({ powerLoading: true })
+      const data = {
+        chatbotId: this.state.chatbot._id,
+        published: true
+      }
+      this.props.updateChatbot(data, (res) => this.afterPublish(res))
+    }
+  }
+
+  afterPublish(res) {
+    if (res.status === 'success') {
+      this.msg.success('Chatbot published successfully')
+      let chatbot = this.state.chatbot
+      chatbot.published = true
+      this.setState({ chatbot, powerLoading: false })
+    } else {
+      this.msg.error(res.description)
+    }
+  }
+
+  showTestModal() {
+    if (!this.props.superUser) {
+      this.setState({ showTestContent: true }, () => {
+        this.refs._open_test_chatbot_modal.click()
+      })
+    } else {
+      this.msg.error('You are not allowed to perform this action')
+    }
+  }
+
+  toggleTestModalContent() {
+    this.setState({ showTestContent: !this.state.showTestContent })
+  }
+
+  getTestModalContent() {
+    if (this.state.showTestContent) {
+      return (
+        <MessengerPlugin
+          appId={this.props.fbAppId}
+          pageId={this.state.chatbot.pageFbId}
+          size='large'
+          passthroughParams='_chatbot'
+        />
+      )
+    } else {
+      return (<div />)
+    }
+  }
+
+  showTreeStructure() {
+    this.setState({ showTreeStructure: true })
+  }
+
+  closeTreeStructure() {
+    this.setState({ showTreeStructure: false })
+  }
+
+  disableShopifyChatbot() {
+    this.props.updateChatbot({
+      chatbotId: this.state.existingChatbot._id,
+      published: false
+    }, (res) => {
+      if (res.status === 'success') {
+        let existingChatbot = this.state.existingChatbot
+        existingChatbot.published = false
+        this.setState({ existingChatbot }, () => {
+          this.onPublish()
+        })
+      } else {
+        this.msg.error(res.description)
+      }
+    })
+  }
+
+  switchBuilder () {
+    let selectedText = 'Advanced Builder'
+    const data = {
+      chatbotId: this.state.chatbot._id,
+      builderPreference: 'advanced'
+    }
+    this.props.updateChatbot(data, (res) => {
+      if (res.status === 'success') {
+        this.setState({builder: 'advanced', selectedText})
+      } else {
+        this.msg.error(res.description || 'Failed to set builder preference')
+      }
+    })
   }
 
   componentWillUnmount () {
@@ -268,6 +413,105 @@ class ConfigureChatbot extends React.Component {
       <div className='m-grid__item m-grid__item--fluid m-wrapper'>
         <AlertContainer ref={(a) => { this.msg = a }} {...alertOptions} />
 
+        <div className='m-subheader '>
+          <div className='d-flex align-items-center'>
+            <div className='mr-auto'>
+              <h3 className='m-subheader__title'>Configure Chatbot - {this.props.location.state.page.pageName}</h3>
+            </div>
+            <div className='row'>
+              {
+                this.state.builder === 'basic' &&
+                <ReactTooltip
+                  id='tree-tructure'
+                  place='bottom'
+                  type='dark'
+                  effect='solid'
+                />
+              }
+              {
+                this.state.builder === 'basic' &&
+                <button
+                  id='_chatbot_message_area_header_analytics'
+                  style={{ marginLeft: '10px', borderColor: '#ffb822' }}
+                  className="pull-right btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill m-btn--air"
+                  onClick={this.showTreeStructure}
+                  data-tip='View Tree Structure'
+                  data-for='tree-tructure'
+                >
+                  <i className="fa flaticon-network"></i>
+                </button>
+              }
+              <ReactTooltip
+                id='chatbot-analytics'
+                place='bottom'
+                type='dark'
+                effect='solid'
+              />
+              <button
+                id='_chatbot_message_area_header_analytics'
+                style={{ marginLeft: '10px', borderColor: '#36a3f7' }}
+                className="pull-right btn btn-info m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill m-btn--air"
+                onClick={this.onAnalytics}
+                data-tip='Chatbot Analytics'
+                data-for='chatbot-analytics'
+              >
+                <i className="fa flaticon-analytics"></i>
+              </button>
+              <ReactTooltip
+                id='test-chatbot'
+                place='bottom'
+                type='dark'
+                effect='solid'
+              />
+              <button
+                id='_chatbot_message_area_header_test'
+                style={{ marginLeft: '10px', borderColor: '#716aca' }}
+                className="pull-right btn btn-brand m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill m-btn--air"
+                onClick={this.showTestModal}
+                data-tip='Test Chatbot'
+                data-for='test-chatbot'
+                disabled={this.state.progress !== 100}
+              >
+                <i className="fa flaticon-paper-plane"></i>
+              </button>
+              <ReactTooltip
+                id='publish-chatbot'
+                place='bottom'
+                type='dark'
+                effect='solid'
+              />
+              <button
+                id='_chatbot_message_area_header_publish'
+                style={{ margin: '0px 10px', borderColor: this.state.chatbot.published ? '#34bfa3' : '#f4516c' }}
+                className={`pull-right btn btn-${this.state.chatbot.published ? 'success' : 'danger'} m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill m-btn--air`}
+                onClick={this.state.chatbot.published ? this.onDisable : this.onPublish}
+                data-tip={this.state.chatbot.published ? 'Disable Chatbot' : 'Publish Chatbot'}
+                data-for='publish-chatbot'
+                disabled={this.state.progress !== 100}
+              >
+                {
+                  this.state.powerLoading
+                    ? <i className="m-loader" />
+                    : <i className="la la-power-off" />
+                }
+              </button>
+              {
+                this.state.builder === 'basic' &&
+                <DROPDOWN
+                  id='chatbot-builder'
+                  selectedText={this.state.builderText}
+                  builder={this.state.builder}
+                  colorStyle='brand'
+                  options={[
+                    {action: () => {}, text: 'Basic Builder', value: 'basic'},
+                    {action: () => {this.switchBuilderTrigger.click()}, text: 'Advanced Builder', value: 'advanced'}
+                  ]}
+                />
+              }
+            </div>
+          </div>
+        </div>
+
         <Prompt
           when={this.state.unsavedChanges}
           message='You have unsaved changes, are you sure you want to leave?'
@@ -280,47 +524,54 @@ class ConfigureChatbot extends React.Component {
             <span className='m--font-brand'>Please wait...</span>
           </div>
           : <div id='_chatbot_main_container'>
-            <div id='_chatbot_configure_area' style={{margin: '15px'}} className='row'>
-              <SIDEBAR
-                data={this.state.sidebarItems}
-                currentBlock={this.state.currentBlock}
+            {
+              this.state.builder === 'basic'
+              ? <BASICBUILDER
                 blocks={this.state.blocks}
-                updateParentState={this.updateState}
                 chatbot={this.state.chatbot}
+                sidebarItems={this.state.sidebarItems}
+                currentBlock={this.state.currentBlock}
+                progress={this.state.progress}
+                unsavedChanges={this.state.unsavedChanges}
+                attachmentUploading={this.state.attachmentUploading}
+                allTriggers={this.state.allTriggers}
+                updateParentState={this.updateState}
+                handleMessageBlock={this.props.handleMessageBlock}
+                user={this.props.user}
+                uploadAttachment={this.props.uploadAttachment}
+                handleAttachment={this.props.handleAttachment}
+                updateChatbot={this.props.updateChatbot}
+                deleteMessageBlock={this.props.deleteMessageBlock}
+                checkWhitelistedDomains={this.props.checkWhitelistedDomains}
+                toggleWhitelistModal={this.toggleWhitelistModal}
+                alertMsg={this.msg}
+              />
+              : <ADVANCEDBUILDER
+                blocks={this.state.blocks}
+                chatbot={this.state.chatbot}
+                sidebarItems={this.state.sidebarItems}
+                currentBlock={this.state.currentBlock}
+                progress={this.state.progress}
+                unsavedChanges={this.state.unsavedChanges}
+                attachmentUploading={this.state.attachmentUploading}
+                allTriggers={this.state.allTriggers}
+                updateParentState={this.updateState}
                 fetchBackup={this.props.fetchBackup}
                 createBackup={this.props.createBackup}
                 restoreBackup={this.props.restoreBackup}
-                alertMsg={this.msg}
                 fetchChatbotDetails={this.fetchChatbotDetails}
                 fetchChatbot={this.fetchChatbot}
-                unsavedChanges={this.state.unsavedChanges}
                 handleMessageBlock={this.props.handleMessageBlock}
-              />
-              <MESSAGEAREA
-                block={this.state.currentBlock}
-                chatbot={this.state.chatbot}
-                alertMsg={this.msg}
+                superUser={this.props.superUser}
                 uploadAttachment={this.props.uploadAttachment}
                 handleAttachment={this.props.handleAttachment}
-                currentLevel={this.state.currentLevel}
-                maxLevel={this.state.chatbot.maxLevels}
-                blocks={this.state.blocks}
-                handleMessageBlock={this.props.handleMessageBlock}
-                fbAppId={this.props.fbAppId}
-                changeActiveStatus={this.props.updateChatbot}
+                updateChatbot={this.props.updateChatbot}
                 deleteMessageBlock={this.props.deleteMessageBlock}
-                registerSocketAction={registerAction}
-                progress={this.state.progress}
-                updateParentState={this.updateState}
-                sidebarItems={this.state.sidebarItems}
                 checkWhitelistedDomains={this.props.checkWhitelistedDomains}
                 toggleWhitelistModal={this.toggleWhitelistModal}
-                onAnalytics={this.onAnalytics}
-                allTriggers={this.state.allTriggers}
-                attachmentUploading={this.state.attachmentUploading}
-                superUser={this.props.superUser}
+                alertMsg={this.msg}
               />
-            </div>
+            }
             <PROGRESS progress={`${this.state.progress}%`} />
             <BACKBUTTON
               onBack={this.onBack}
@@ -336,6 +587,27 @@ class ConfigureChatbot extends React.Component {
               title='Manage Whitelist Domains'
               content={this.getWhitelistModalContent()}
             />
+            <button ref='_open_test_chatbot_modal' style={{ display: 'none' }} data-toggle='modal' data-target='#_test_chatbot' />
+            <MODAL
+              id='_test_chatbot'
+              title='Test Chatbot'
+              content={this.getTestModalContent()}
+              onClose={this.toggleTestModalContent}
+            />
+            <button style={{ display: 'none' }} ref={el => this.disableShopifyChatbotTrigger = el} data-toggle='modal' data-target='#disableShopifyChatbot' />
+            <CONFIRMATIONMODAL
+              id='disableShopifyChatbot'
+              title='Disable Shopify Chatbot'
+              description='You have a Shopify chatbot already enabled for this page. By enabling this manual chatbot, that Shopify chatbot will be disabled. Do you wish to continue?'
+              onConfirm={this.disableShopifyChatbot}
+            />
+            <button style={{ display: 'none' }} ref={el => this.switchBuilderTrigger = el} data-toggle='modal' data-target='#switchBuilder' />
+            <CONFIRMATIONMODAL
+              id='switchBuilder'
+              title='Warning'
+              description="Please note that if you set the builder preference to advanced, then you won't be able to switch back to basic. Are you sure you want to continue?"
+              onConfirm={this.switchBuilder}
+            />
           </div>
         }
       </div>
@@ -346,6 +618,7 @@ class ConfigureChatbot extends React.Component {
 function mapStateToProps (state) {
   return {
     fbAppId: state.basicInfo.fbAppId,
+    user: (state.basicInfo.user),
     superUser: (state.basicInfo.superUser)
   }
 }
